:broom: moved Landing into separated component
